# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Book Favorites API
  description: A REST API for managing book collections, user favorites, reading lists, and reviews
  version: 1.0.0
  contact:
    name: API Support
    email: support@bookfavorites.com
servers:
  - url: http://localhost:4000/api/v1
    description: Development server
  - url: http://localhost:4000/api
    description: Development server (legacy, redirects to v1)
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
  schemas:
...
    Book:
      type: object
      properties:
        id:
          type: string
          example: "1"
        title:
          type: string
          example: "To Kill a Mockingbird"
        author:
          type: string
          example: "Harper Lee"
        genre:
          type: string
          example: "Fiction"
    Review:
      type: object
      properties:
        id:
          type: string
          example: "1774478036832"
        bookId:
          type: string
          example: "1"
        username:
          type: string
          example: "user1"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: "A masterpiece"
        createdAt:
          type: string
          format: date-time
          example: "2026-03-25T22:33:56.832Z"
...
    CreateReviewRequest:
      type: object
      required:
        - bookId
        - rating
      properties:
        bookId:
          type: string
          example: "1"
        rating:
          type: integer
          minimum: 1
          maximum: 5
          example: 5
        comment:
          type: string
          example: "A masterpiece"
...
    AddWantToReadRequest:
      type: object
      required:
        - bookId
      properties:
        bookId:
          type: string
          example: "1"
    SuccessMessage:
      type: object
      properties:
        message:
          type: string
          example: "Operation completed successfully"
    ErrorResponse:
      type: object
      properties:
        error:
          type: object
          properties:
            code:
              type: string
              example: "VALIDATION_ERROR"
            message:
              type: string
              example: "Invalid input data"
...
paths:
...
  /reviews/{bookId}:
    get:
      summary: Get all reviews for a specific book
      tags:
        - Reviews
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
          description: ID of the book
      responses:
        '200':
          description: List of reviews for the book
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Review'
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                book_not_found:
                  value:
                    error:
                      code: "BOOK_NOT_FOUND"
                      message: "Book not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reviews:
    post:
      summary: Create a new review for a book
      tags:
        - Reviews
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateReviewRequest'
      responses:
        '201':
          description: Review created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Review'
        '400':
          description: Invalid input data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  value:
                    error:
                      code: "VALIDATION_ERROR"
                      message: "bookId and rating are required"
        '401':
          description: Authentication required
        '404':
          description: Book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /reviews/{reviewId}:
    delete:
      summary: Delete own review
      tags:
        - Reviews
      security:
        - bearerAuth: []
      parameters:
        - name: reviewId
          in: path
          required: true
          schema:
            type: string
          description: ID of the review to delete
      responses:
        '204':
          description: Review deleted successfully
        '401':
          description: Authentication required
        '403':
          description: Not authorized to delete this review
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                forbidden:
                  value:
                    error:
                      code: "FORBIDDEN"
                      message: "Not authorized to delete this review"
        '404':
          description: Review not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                review_not_found:
                  value:
                    error:
                      code: "REVIEW_NOT_FOUND"
                      message: "Review not found"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /want-to-read:
    get:
      summary: Get user's want-to-read list
      tags:
        - Want to Read
      security:
        - bearerAuth: []
      responses:
        '200':
          description: List of books in user's want-to-read list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Book'
        '401':
          description: Authentication required
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

    post:
      summary: Add book to want-to-read list
      tags:
        - Want to Read
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AddWantToReadRequest'
      responses:
        '201':
          description: Book added to want-to-read list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessMessage'
        '400':
          description: Invalid book ID or book not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                validation_error:
                  value:
                    error:
                      code: "VALIDATION_ERROR"
                      message: "Valid Book ID is required"
                book_not_found:
                  value:
                    error:
                      code: "BOOK_NOT_FOUND"
                      message: "Book not found"
        '401':
          description: Authentication required
        '404':
          description: User not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /want-to-read/{bookId}:
    delete:
      summary: Remove book from want-to-read list
      tags:
        - Want to Read
      security:
        - bearerAuth: []
      parameters:
        - name: bookId
          in: path
          required: true
          schema:
            type: string
          description: ID of the book to remove
      responses:
        '204':
          description: Book removed from want-to-read list
        '401':
          description: Authentication required
        '404':
          description: User not found or book not in want-to-read list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              examples:
                user_not_found:
                  value:
                    error:
                      code: "USER_NOT_FOUND"
                      message: "User not found"
                book_not_in_list:
                  value:
                    error:
                      code: "BOOK_NOT_IN_LIST"
                      message: "Book not in want-to-read list"
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...