# backend/routes/wantToRead.js
const express = require('express');

function createWantToReadRouter({ usersFile, booksFile, readJSON, writeJSON, authenticateToken }) {
  const router = express.Router();

  // generated-by-copilot: GET /want-to-read - fetch user's want-to-read books
  router.get('/', authenticateToken, (req, res) => {
    try {
      const users = readJSON(usersFile);
      const user = users.find(u => u.username === req.user.username);
      if (!user) return res.status(404).json({ error: { code: 'USER_NOT_FOUND', message: 'User not found' } });
      
      // generated-by-copilot: ensure wantToRead array exists and default to empty array
      const wantToReadIds = user.wantToRead || [];
      
      const books = readJSON(booksFile);
      const wantToReadBooks = books.filter(b => wantToReadIds.indexOf(b.id) !== -1);
      res.json(wantToReadBooks);
    } catch {
      res.status(500).json({ error: { code: 'SERVER_ERROR', message: 'An unexpected error occurred.' } });
    }
  });

  // generated-by-copilot: POST /want-to-read - add a book to user's want-to-read list
  router.post('/', authenticateToken, (req, res) => {
    const { bookId } = req.body;
    if (!bookId || typeof bookId !== 'string') return res.status(400).json({ error: { code: 'VALIDATION_ERROR', message: 'Valid Book ID is required' } });
    
    try {
      const users = readJSON(usersFile);
      const user = users.find(u => u.username === req.user.username);
      if (!user) return res.status(404).json({ error: { code: 'USER_NOT_FOUND', message: 'User not found' } });
      
      // generated-by-copilot: validate that the book exists
      const books = readJSON(booksFile);
      const bookExists = books.find(b => b.id === bookId);
      if (!bookExists) return res.status(400).json({ error: { code: 'BOOK_NOT_FOUND', message: 'Book not found' } });
      
      // generated-by-copilot: ensure wantToRead array exists and default to empty array
      if (!user.wantToRead) {
        user.wantToRead = [];
      }
      
      // generated-by-copilot: only add if not already in want-to-read list
      if (user.wantToRead.indexOf(bookId) === -1) {
        user.wantToRead.push(bookId);
        writeJSON(usersFile, users);
      }
      res.status(201).json({ message: 'Book added to want-to-read list successfully' });
    } catch {
      res.status(500).json({ error: { code: 'SERVER_ERROR', message: 'An unexpected error occurred.' } });
    }
  });

  // generated-by-copilot: DELETE /want-to-read/:bookId - remove a book from user's want-to-read list
  router.delete('/:bookId', authenticateToken, (req, res) => {
    const { bookId } = req.params;
    try {
      const users = readJSON(usersFile);
      const user = users.find(u => u.username === req.user.username);
      if (!user) return res.status(404).json({ error: { code: 'USER_NOT_FOUND', message: 'User not found' } });
      
      // generated-by-copilot: ensure wantToRead array exists and default to empty array
      if (!user.wantToRead) {
        user.wantToRead = [];
      }
      
      const index = user.wantToRead.indexOf(bookId);
      if (index === -1) return res.status(404).json({ error: { code: 'BOOK_NOT_IN_LIST', message: 'Book not in want-to-read list' } });
      
      user.wantToRead.splice(index, 1);
      writeJSON(usersFile, users);
      res.status(204).send();
    } catch {
      res.status(500).json({ error: { code: 'SERVER_ERROR', message: 'An unexpected error occurred.' } });
    }
  });

  return router;
}

module.exports = createWantToReadRouter;